qr: print_oops.c: bug fix + qr multi-part message
Since we don't have any way of knowing if what we write on the
framebuffer actually gets to the screen and there is no simple way to
guarantee that what we draw is always on top, we are going to make a
kthread and continuously draw the QR on the framebuffer. This approach
is mandatory anyway since we are going to draw an animation with multiple
QRs.

Fixed a bug in compress that would exit with error if the compressed size
is bigger than the uncompressed size.

qr: print_oops.c: add BK1 multi-part message format + fallback to ASCII

Implement Bug Koops Android application format for multi-part messages.
For more informations check github.com/links234/BugKoops

If there is no available framebuffer just print QRs in ASCII art!

qr: print_oops.c: add sysfs parameters

For BugKoops to QR module communication we have to use the user (tell him
what to do). For user to QR module we will use sysfs.

qr: print_oops.c: implement tortoise and rabbit strategy

This is the best strategy that I found for 1-way data transfer and it's
fairly simple. The basic ideea is to iterate circular with 2 pointers
one being twice as fastest as the other. For more information please
check docs/FORMAT.md at github.com/links234/BugKoops