qr: print_oops.c: various bug fixes
The console will overwrite the framebuffer, but only where it really has
to draw, so for the most time there will still remain unwanted parts of
the previous QR code that was drawn to the screen. Fix this by drawing a
black rectangle over the last QR.

qr: print_oops.c: fix ASCII fallback

If there are no framebuffer it will still exit before reaching the
while.